+++ __tests__/car.test.js
@@ -1,4 +1,4 @@
-import Car from "../src/car.js";
+import Car from "../src/domain/car.js";
 
 describe("Car Class 테스트", () => {
     let car;
