+++ __tests__/race.test.js
@@ -1,5 +1,5 @@
-import Race from "../src/race.js";
-import Car from "../src/car.js";
+import Race from "../src/domain/race.js";
+import Car from "../src/domain/car.js";
 
 describe("Race Class 테스트", () => {
     let race;
